+++ CHANGELOG.rst
@@ -1,8 +1,9 @@
 1.2.3
 -----
 
-* :meth:`.Table.distinct` now accepts sequences of column names as a key.
-* :meth:`.Table.join` now accepts sequences of column names as either a left or right key. (#475)
+* :meth:`.Table.order_by` now accepts a sequence of column names as a key.
+* :meth:`.Table.distinct` now accepts a sequence of column names as a key.
+* :meth:`.Table.join` now accepts a sequence of column names as either a left or right key. (#475)
 * Cookbook: Create a table from a DBF file. (#472)
 * Cookbook: Create a table from an Excel spreadsheet.
 * Added explicit error if a filename is passed to the :class:`.Table` constructor. (#438)